socket instnace global resource, moved scoket stuff to package
