--- src/websocket.py
@@ -1,12 +1,0 @@
-from tornado.websocket import WebSocketHandler
-
-class WebSocket(WebSocketHandler):
-    def open(self):
-        print("Socket opened.")
-
-    def on_message(self, message):
-        self.write_message("Received: " + message)
-        print("Received message: " + message)
-
-    def on_close(self):
-        print("Socket closed.")
